feat(runtime): export built-in presets
