+++ docs/integrations/runtime.md
@@ -47,14 +47,6 @@ The runtime does not come with preflights, if you want to have style resets, you
 
 Several builds are available for different use cases.
 
-### Core
-
-Without any preset:
-
-```html
-<script src="https://cdn.jsdelivr.net/npm/@unocss/runtime/core.global.js"></script>
-```
-
 ### Uno (default)
 
 With `@unocss/preset-uno` preset:
@@ -77,6 +69,25 @@ With `@unocss/preset-mini` and `@unocss/preset-attributify` preset:
 
 ```html
 <script src="https://cdn.jsdelivr.net/npm/@unocss/runtime/mini.global.js"></script>
+```
+
+### Core
+
+If you need to mix and match presets, you can load only the core runtime and assign the presets manually. All the [official presets](/presets/#presets) from UnoCSS are available and must be loaded before initializing the core runtime.
+
+```html
+<script src="https://cdn.jsdelivr.net/npm/@unocss/runtime/preset-icons.global.js"></script>
+<script>
+  window.__unocss = {
+    presets: [
+      () => window.__unocss_runtime.presets.presetIcons({
+        scale: 1.2,
+        cdn: 'https://esm.sh/'
+      }),
+    ],
+  }
+</script>
+<script src="https://cdn.jsdelivr.net/npm/@unocss/runtime/core.global.js"></script>
 ```
 
 ## Bundler Usage
