+++ packages/runtime/src/presets/preset-mini.ts
@@ -1,0 +1,7 @@
+import type { PresetMiniOptions } from '@unocss/preset-mini'
+import presetMini from '@unocss/preset-mini'
+
+window.__unocss_runtime = window.__unocss_runtime ?? { presets: {} }
+window.__unocss_runtime.presets = Object.assign(window.__unocss_runtime?.presets ?? {}, {
+  presetMini: (options: PresetMiniOptions) => presetMini(options),
+})
