+++ packages/runtime/src/presets/preset-uno.ts
@@ -1,0 +1,7 @@
+import type { PresetUnoOptions } from '@unocss/preset-uno'
+import presetUno from '@unocss/preset-uno'
+
+window.__unocss_runtime = window.__unocss_runtime ?? { presets: {} }
+window.__unocss_runtime.presets = Object.assign(window.__unocss_runtime?.presets ?? {}, {
+  presetUno: (options: PresetUnoOptions) => presetUno(options),
+})
