+++ packages/runtime/src/presets/preset-wind.ts
@@ -1,0 +1,7 @@
+import type { PresetWindOptions } from '@unocss/preset-wind'
+import presetWind from '@unocss/preset-wind'
+
+window.__unocss_runtime = window.__unocss_runtime ?? { presets: {} }
+window.__unocss_runtime.presets = Object.assign(window.__unocss_runtime?.presets ?? {}, {
+  presetWind: (options: PresetWindOptions) => presetWind(options),
+})
